requirements.txt
stravalib
python-dateutil
loguru
tqdm
fire